+++ packages/prover-client/src/effect/WasmProver.ts
@@ -167,7 +167,7 @@ class WasmProverImpl implements Context.Tag.Service<ProverClient> {
         op: 'prove',
         args: [serializedPreimage, overwriteBindingInput],
       }),
-    lookupKey: async (keyLocation: string): Promise<ProvingKeyMaterial | undefined> =>
+    lookupKey: (keyLocation: string): Promise<ProvingKeyMaterial | undefined> =>
       (keyMaterialProvider ?? this.keyMaterialProvider).lookupKey(keyLocation),
   });