+++ packages/unshielded-wallet/src/index.ts
@@ -11,6 +11,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 export * from './UnshieldedWallet.js';
-export { type UnshieldedTransactionHistoryEntry, UnshieldedSectionSchema } from './v1/TransactionHistory.js';
+export {
+  type UnshieldedTransactionHistoryEntry,
+  UnshieldedSectionSchema,
+  mergeUnshieldedSections,
+} from './v1/TransactionHistory.js';
 export { type SignSegment } from './v1/Signing.js';
 export * from './KeyStore.js';